OC-104: refactoring, wallet address regeneration
