+++ core/consts/currencies.py
@@ -4,6 +4,12 @@ from collections.abc import Callable
 
 from core.currency import Currency, TokenParams, CoinParams
 
+@dataclass
+class BlockchainAccount:
+    address: str
+    private_key: str
+    redeem_script: Optional[str] = None
+
 ALL_CURRENCIES: List[Currency] = []  # all Currency instances
 
 CURRENCIES_LIST: List[Tuple[int, str]] = []
@@ -25,11 +31,4 @@ CRYPTO_WALLET_CREATORS: Union[Dict[Currency, Callable], Dict[Currency, Dict[str,
 
 CRYPTO_COINS_PARAMS: Dict[Currency, CoinParams] = {}
 
-CRYPTO_WALLET_ACCOUNT_CREATORS = {}
-
-
-@dataclass
-class WalletAccount:
-    address: str
-    private_key: str
-    redeem_script: Optional[str] = None
+CRYPTO_WALLET_ACCOUNT_CREATORS: Dict[Currency, BlockchainAccount] = {}
