+++ cryptocoins/coins/trx/wallet.py
@@ -4,9 +4,32 @@ from django.conf import settings
 from django.db import transaction
 from tronpy.keys import PrivateKey
 
+from core.consts.currencies import BlockchainAccount
 from lib.cipher import AESCoderDecoder
 
 log = logging.getLogger(__name__)
+
+
+def create_trx_address():
+    while 1:
+        account = PrivateKey.random()
+        private_key = account.hex()
+        address = account.public_key.to_base58check_address()
+
+        encrypted_key = AESCoderDecoder(settings.CRYPTO_KEY).encrypt(private_key)
+        decrypted_key = AESCoderDecoder(settings.CRYPTO_KEY).decrypt(encrypted_key)
+
+        if decrypted_key == private_key:
+            break
+    return address, encrypted_key
+
+
+def create_new_blockchain_account() -> BlockchainAccount:
+    address, encrypted_pk = create_trx_address()
+    return BlockchainAccount(
+        address=address,
+        private_key=AESCoderDecoder(settings.CRYPTO_KEY).decrypt(encrypted_pk),
+    )
 
 
 @transaction.atomic
@@ -27,22 +50,13 @@ def get_or_create_trx_wallet(user_id, is_new=False):
     if not is_new and user_wallet is not None:
         return user_wallet
 
-    while 1:
-        account = PrivateKey.random()
-        private_key = account.hex()
-        address = account.public_key.to_base58check_address()
-
-        encrypted_key = AESCoderDecoder(settings.CRYPTO_KEY).encrypt(private_key)
-        decrypted_key = AESCoderDecoder(settings.CRYPTO_KEY).decrypt(encrypted_key)
-
-        if decrypted_key == private_key:
-            break
+    address, encrypted_key = create_trx_address()
 
     user_wallet = UserWallet.objects.create(
         user_id=user_id,
-        currency=TRX_CURRENCY,
         address=address,
         private_key=encrypted_key,
+        currency=TRX_CURRENCY,
         blockchain_currency=TRX_CURRENCY
     )
 
@@ -65,11 +79,13 @@ def get_or_create_trc20_wallet(user_id, currency, is_new=False):
     if not is_new and trc20_wallet is not None:
         return trc20_wallet
 
+    address, encrypted_key = create_trx_address()
+
     trc20_wallet = UserWallet.objects.create(
         user_id=user_id,
+        address=address,
+        private_key=encrypted_key,
         currency=currency,
-        address=trx_wallet.address,
-        private_key=trx_wallet.private_key,
         blockchain_currency=TRX_CURRENCY
     )
